Fix bug when rebasing a checked out branch with conflicts
If a branch was checked out in a worktree we try and rebase it there. If the rebase command fails we need to check if it is still rebasing and therefore failed due to conflicts. This would look in the wrong place previously, which would mean the rebase wasn't aborted properly and then future stuff would fail.

This fixes it as far as I can tell. I added a test.